[CI] Add a CI job to run type checking with mypy
Fixed type checking issues on the whole codebase.

+++ src/aleph/model/hashes.py
@@ -33,6 +33,8 @@ from gridfs.errors import NoFile
 
 async def get_value(key: str):
     from aleph.model import fs
+    if fs is None:
+        raise ValueError("MongoDB fs not initialized")
 
     try:
         gridout = await fs.open_download_stream_by_name(key)
@@ -43,6 +45,8 @@ async def get_value(key: str):
 
 async def set_value(key: str, value: bytes):
     from aleph.model import fs
+    if fs is None:
+        raise ValueError("MongoDB fs not initialized")
 
     file_id = await fs.upload_from_stream(key, value)
     return file_id
@@ -50,6 +54,8 @@ async def set_value(key: str, value: bytes):
 
 async def delete_value(key: str):
     from aleph.model import fs
+    if fs is None:
+        raise ValueError("MongoDB fs not initialized")
 
     async for gridfs_file in fs.find({"filename": key}):
         await fs.delete(gridfs_file._id)
